requirements.txt
openai==1.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
streamlit==1.28.2
tika==2.6.0